feature: Implement initial movie detail component
Add Hashed router to navigate between react page urls

change react page url to /app
