+++ src/commonMain/resources/index.html
@@ -6,7 +6,6 @@
 </head>
 <body>
     <div id="root"></div>
-    aaaaaaaa
     <script src="movieratingapp.js"></script>
 </body>
 </html>
